Pulse-Guard/backend/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
google-cloud-storage
google-genai
python-multipart
pydantic
python-dotenv
